+++ sdk/src/main/java/com/horizen/tools/utils/MessagePrinter.java
@@ -1,0 +1,5 @@
+package com.horizen.tools.utils;
+
+public interface MessagePrinter {
+    void print(String message);
+}