+++ sdk/src/main/java/com/horizen/wallet/ApplicationWallet.java
@@ -12,4 +12,8 @@ public interface ApplicationWallet {
     void onRemoveSecret(Proposition proposition);
     void onChangeBoxes(byte[] blockId, List<Box<Proposition>> boxesToUpdate, List<byte[]> boxIdsToRemove);
     void onRollback(byte[] blockId);
+
+    // check that all storages of the application which are update by the sdk core, have the version corresponding to the
+    // blockId given. This is useful when checking the alignment of the storages versions at node restart
+    boolean checkStoragesVersion(byte[] blockId);
 }