+++ sdk/src/main/scala/com/horizen/SidechainState.scala
@@ -18,7 +18,7 @@ import com.horizen.utils.{BlockFeeInfo, ByteArrayWrapper, BytesUtils, FeePayment
 import scorex.core._
 import scorex.core.transaction.state.{BoxStateChangeOperation, BoxStateChanges, Insertion, MinimalState, ModifierValidation, Removal, TransactionValidation}
 import scorex.crypto.hash.Blake2b256
-import scorex.util.{ModifierId, ScorexLogging}
+import scorex.util.{ModifierId, ScorexLogging, bytesToId}
 
 import java.math.{BigDecimal, MathContext}
 import com.horizen.box.data.ZenBoxData
@@ -44,8 +44,6 @@ class SidechainState private[horizen] (stateStorage: SidechainStateStorage,
     with UtxoMerkleTreeView
 {
 
-  checkVersion()
-
   override type NVCT = SidechainState
 
   lazy val verificationKeyFullFilePath: String = {
@@ -61,34 +59,6 @@ class SidechainState private[horizen] (stateStorage: SidechainStateStorage,
       log.info(s"Verification key file at location: ${verificationFile.getAbsolutePath}")
       verificationFile.getAbsolutePath
     }
-  }
-
-  private def checkVersion(): Unit = {
-    val versionBytes = versionToBytes(version)
-
-    require({
-      stateStorage.lastVersionId match {
-        case Some(storageVersion) => storageVersion.data.sameElements(versionBytes)
-        case None => true
-      }
-    },
-      s"Specified version is invalid. StateStorage version ${stateStorage.lastVersionId.map(w => bytesToVersion(w.data)).getOrElse(version)} != $version")
-
-    require({
-      forgerBoxStorage.lastVersionId match {
-        case Some(storageVersion) => storageVersion.data.sameElements(versionBytes)
-        case None => true
-      }
-    },
-      s"Specified version is invalid. StateForgerBoxStorage version ${forgerBoxStorage.lastVersionId.map(w => bytesToVersion(w.data)).getOrElse(version)} != $version")
-
-    require({
-      utxoMerkleTreeStorage.lastVersionId match {
-        case Some(storageVersion) => storageVersion.data.sameElements(versionBytes)
-        case None => true
-      }
-    },
-      s"Specified version is invalid. UtxoMerkleTreeStorage version ${utxoMerkleTreeStorage.lastVersionId.map(w => bytesToVersion(w.data)).getOrElse(version)} != $version")
   }
 
   // Note: emit tx.semanticValidity for each tx
@@ -390,7 +360,10 @@ class SidechainState private[horizen] (stateStorage: SidechainStateStorage,
           newVersion,
           appState
         )
-      case Failure(exception) => throw exception
+      case Failure(exception) => {
+        log.error("call to onApplyChanges() method has failed: ", exception)
+        throw exception
+      }
     }
   }.recoverWith{
     case exception =>
@@ -404,16 +377,28 @@ class SidechainState private[horizen] (stateStorage: SidechainStateStorage,
 
   override def rollbackTo(to: VersionTag): Try[SidechainState] = Try {
     require(to != null, "Version to rollback to must be NOT NULL.")
+    log.debug(s"rolling back state to version = ${to}")
     val version = BytesUtils.fromHexString(to)
+    val bawVersion = new ByteArrayWrapper(version)
+
+    val forgerBoxStorageNew = forgerBoxStorage.rollback(bawVersion).get
+    val stateStorageNew = stateStorage.rollback(bawVersion).get
+    val utxoMerkleTreeStorageNew = utxoMerkleTreeStorage.rollback(bawVersion).get
+
     applicationState.onRollback(version) match {
-      case Success(appState) => new SidechainState(
-        stateStorage.rollback(new ByteArrayWrapper(version)).get,
-        forgerBoxStorage.rollback(new ByteArrayWrapper(version)).get,
-        utxoMerkleTreeStorage.rollback(new ByteArrayWrapper(version)).get,
-        params,
-        to,
-        appState)
-      case Failure(exception) => throw exception
+      case Success(appState) => {
+        new SidechainState(
+          stateStorageNew,
+          forgerBoxStorageNew,
+          utxoMerkleTreeStorageNew,
+          params,
+          to,
+          appState)
+      }
+      case Failure(exception) => {
+        log.error("call to applicationState.onRollback() method has failed: ", exception)
+        throw exception
+      }
     }
   }.recoverWith{case exception =>
     log.error("Exception was thrown during rollback.", exception)
@@ -458,6 +443,46 @@ class SidechainState private[horizen] (stateStorage: SidechainStateStorage,
   // Check that State is on the last index of the withdrawal epoch: last block applied have finished the epoch.
   def isWithdrawalEpochLastIndex: Boolean = {
     WithdrawalEpochUtils.isEpochLastIndex(stateStorage.getWithdrawalEpochInfo.getOrElse(WithdrawalEpochInfo(0,0)), params)
+  }
+
+  // Check that all storages are consistent and in case try some rollbacks.
+  // Return the state and common version, throw an exception if some unrecoverable misalignment has been detected
+  def ensureStorageConsistencyAfterRestore: Try[(SidechainState)] = Try {
+    // updates are in order:
+    //      appState--> utxoMerkleTreeStorage --> stateStorage --> forgerBoxStorage
+
+    // get the version of the last updated storage and check that the others have the same
+    // version
+    val versionFbBytes = forgerBoxStorage.lastVersionId.get.data()
+    val appStateVersionOk = applicationState.checkStoragesVersion(versionFbBytes)
+
+    val versionFb = bytesToId(versionFbBytes)
+
+    if (appStateVersionOk) {
+      // appState is aligned with the last storage version, require that also intermediate storages have the same version
+
+      // TODO csw capability will be optional, related utxo mrkl tree storage might even be empty
+      if (utxoMerkleTreeStorage.lastVersionId.isDefined) {
+        val versionUmt = bytesToId(utxoMerkleTreeStorage.lastVersionId.get.data())
+        require(versionFb == versionUmt, "ForgerBox and Utxo storage versions must be aligned")
+      }
+
+      val versionSt  = bytesToId(stateStorage.lastVersionId.get.data())
+      require(versionFb == versionSt, "ForgerBox and State storage versions must be aligned")
+      require(versionFb == versionToId(version), "ForgerBox version and SidechainState version attribute must be aligned")
+      log.debug("All state storages are consistent")
+
+      this
+    } else {
+      log.debug("state storages are not consistent")
+
+      val rolledBackState = rollbackTo(idToVersion(versionFb))
+      if (rolledBackState.isFailure) {
+        throw new IllegalStateException("Could not rollback state")
+      } else {
+        rolledBackState.get
+      }
+    }
   }
 
   // Collect Fee payments while appending the last withdrawal epoch block (optional), considering that block fee info as well.
@@ -519,7 +544,7 @@ object SidechainState
         (sr ++ tx.unlockers().asScala.map(_.closedBoxId()), sa ++ tx.newBoxes().asScala, f + tx.fee())
       }
 
-    // calculate list of ID of unlokers' boxes -> toRemove
+    // calculate list of ID of unlockers' boxes -> toRemove
     // calculate list of new boxes -> toAppend
     // calculate the rewards for Miner/Forger -> create another regular tx OR Forger need to add his Reward during block creation
     @SuppressWarnings(Array("org.wartremover.warts.Product","org.wartremover.warts.Serializable"))