+++ sdk/src/main/scala/com/horizen/consensus/ConsensusDataStorage.scala
@@ -1,8 +1,7 @@
 package com.horizen.consensus
 
 import java.util.{ArrayList => JArrayList}
-
-import com.horizen.storage.Storage
+import com.horizen.storage.{SidechainStorageInfo, Storage}
 import com.horizen.utils.{ByteArrayWrapper, Pair => JPair}
 import scorex.crypto.hash.Blake2b256
 import scorex.util.ScorexLogging
@@ -10,7 +9,9 @@ import scorex.util.ScorexLogging
 import scala.compat.java8.OptionConverters._
 import scala.util.Random
 
-class ConsensusDataStorage(consensusEpochInfoStorage: Storage) extends ScorexLogging {
+class ConsensusDataStorage(consensusEpochInfoStorage: Storage)
+  extends ScorexLogging
+  with SidechainStorageInfo {
   def addStakeConsensusEpochInfo(epochId: ConsensusEpochId, stakeEpochInfo: StakeConsensusEpochInfo): Unit = {
     log.info(s"Storage with id:${this.hashCode()} -- Add stake to consensus data storage: for epochId ${epochId} stake info: ${stakeEpochInfo}")
 
@@ -59,6 +60,12 @@ class ConsensusDataStorage(consensusEpochInfoStorage: Storage) extends ScorexLog
     val listForUpdate = new JArrayList[JPair[ByteArrayWrapper, ByteArrayWrapper]]()
     val addedData = new JPair(key, new ByteArrayWrapper(value))
     listForUpdate.add(addedData)
-    consensusEpochInfoStorage.update(new ByteArrayWrapper(nextVersion), listForUpdate, java.util.Collections.emptyList())
+    val version = new ByteArrayWrapper(nextVersion)
+    consensusEpochInfoStorage.update(version, listForUpdate, java.util.Collections.emptyList())
+    log.debug("Consensus data storage updated with version: " + version)
+  }
+
+  override def lastVersionId: Option[ByteArrayWrapper] = {
+    consensusEpochInfoStorage.lastVersionID().asScala
   }
 }