+++ sdk/src/main/scala/com/horizen/storage/SidechainStorageInfo.scala
@@ -1,0 +1,7 @@
+package com.horizen.storage
+
+import com.horizen.utils.ByteArrayWrapper
+
+trait SidechainStorageInfo {
+  def lastVersionId : Option[ByteArrayWrapper]
+}